# faster `hypot` for `Float32` and `Float16`