--- a/base/Base.jl
+++ b/base/Base.jl
@@ -294,6 +294,9 @@ include("secretbuffer.jl")
 
 # core math functions
 include("floatfuncs.jl")
+# Fast math
+include("fastmath.jl")
+using .FastMath
 include("math.jl")
 using .Math
 const (√)=sqrt
@@ -329,10 +332,6 @@ using .Order
 # Combinatorics
 include("sort.jl")
 using .Sort
-
-# Fast math
-include("fastmath.jl")
-using .FastMath
 
 function deepcopy_internal end