Before:
```julia
julia> @benchmark hypot(x,y) setup=(begin x,y = rand(Float32, 2) end) evals=10000
BenchmarkTools.Trial: 10000 samples with 10000 evaluations.
 Range (min … max):  3.051 ns … 8.185 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     5.700 ns             ┊ GC (median):    0.00%
 Time  (mean ± σ):   5.616 ns ± 0.208 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

                                          ▁▇▇  ▂▇█▅▁▂▂▂     ▂
  ▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁████▇█████████████ █
  3.05 ns     Histogram: log(frequency) by time     6.27 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark hypot(x,y) setup=(begin x,y = rand(Float16, 2) end) evals=10000
BenchmarkTools.Trial: 10000 samples with 10000 evaluations.
 Range (min … max):  3.133 ns … 10.518 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     6.265 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   6.229 ns ±  0.280 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

                                          ▇▆▃▆▁▁█▄▇▂▄▂▁▁▁    ▂
  ▆▅▁▁▁▁▁▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄█████████████████▇ █
  3.13 ns      Histogram: log(frequency) by time     7.06 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

After:
```julia
julia> @benchmark _hypot(x,y) setup=(begin x,y = rand(Float32, 2) end) evals=10000
BenchmarkTools.Trial: 10000 samples with 10000 evaluations.
 Range (min … max):  2.161 ns … 5.114 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     2.188 ns             ┊ GC (median):    0.00%
 Time  (mean ± σ):   2.197 ns ± 0.082 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▂ ▂ ▇█ ▂                                                  ▁
  █▁█▃██▁█▄▃█▃▁▄▃▁▁▁▁▅▃▁▃▁▃▃▄▄▁▁▁▃▁▃▁▄▄▄▃▅▃▄▄▄▄▄▅▃▄▄▆▅▄▄▄▁▅ █
  2.16 ns     Histogram: log(frequency) by time     2.47 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark _hypot(x,y) setup=(begin x,y = rand(Float16, 2) end) evals=10000
BenchmarkTools.Trial: 10000 samples with 10000 evaluations.
 Range (min … max):  2.547 ns … 3.631 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     2.620 ns             ┊ GC (median):    0.00%
 Time  (mean ± σ):   2.630 ns ± 0.070 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

     ▁ ▆ ▄ █ ▆   ▃                                          ▂
  ▇▇▃█▁█▆█▃█▄█▁▇▄██▅▄▅▄▄▃▄▅▅▅▅▄▄▅▅▅▅▆▃▆▅▆▇▇▇▆▆▆▃▅▃▅▅▅▅▄▅▅▄▆ █
  2.55 ns     Histogram: log(frequency) by time        3 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```