@oscardssmith 2021-09-05
I think if you write `@fastmath sqrt(x*x+y*y)`, it will let `x^2+y^2` turn into `fma(x, x, y*y)` which should be faster.