@brett-mahar 2021-09-05
@oscardssmith there was talk of removing fastmath from Julia, apparently its a bit shonky:

https://github.com/JuliaLang/julia/issues/36246