@Moelf 2021-09-05
I think FMA makes it much slower on machine without FMA? don't remember the conclusion