@oscardssmith 2021-09-05
@brett-mahar while we might eventually get rid of it, we won't until there is a good replacement. The reason we still have `@fastmath` is that you need it to get fast code in a bunch of circumstances.
@moelf That's why I recommended using `@fastmath` rather than explicitly writing an fma in. This way, LLVM will be able to use `fma` on the CPUs where there is hardware support, and just do `x*x+y*y` on cpus that don't have fma in hardware.