@Moelf 2021-09-05
hm, fastmath is slower for `Float16` on my computer, 1.6 ns vs. 1.3 ns