@Moelf 2021-09-05
wait nvm, `@fastmath` doesn't exist at this stage