@KristofferC 2021-09-06
> This way, LLVM will be able to use fma on the CPUs where there is hardware support, and just do x*x+y*y on cpus that don't have fma in hardware.

Isn't that just `muladd`? 

I don't think we should add a bunch of `@fastmath` in these type of functions because it is underspecified what transformations `@fastmath` actually allows.