@oscardssmith 2021-09-06
Oh yeah, muladd is probably what we want