@dkarrasch 2021-09-08
Shall we have tests for the Float16 case, including infs and nans? Otherwise this is good to go, if there are no accuracy concerns left.