@KristofferC 2021-09-16
Saw this on Wikipedia on FMA (https://en.wikipedia.org/wiki/Multiply–accumulate_operation#Fused_multiply–add):

> Fused multiply–add can usually be relied on to give more accurate results. However, William Kahan has pointed out that it can give problems if used unthinkingly. If x^2 − y^2 is evaluated as ((x × x) − y × y) (following Kahan's suggested notation in which redundant parentheses direct the compiler to round the (x × x) term first) using fused multiply–add, then the result may be negative even when x = y due to the first multiplication discarding low significance bits. This could then lead to an error if, for instance, the square root of the result is then evaluated.

Here we do that except x^2 + y^2. I guess that's fine then?