@oscardssmith 2021-09-16
The reason this works is because we are doing a higher precision fma. the multiplication of 2 `Float32` has 46 bits of precision or less, so a `Float64` can store `y*y` exactly.